pl - rename checkout.d.ts to hosted-checkout.d.ts
