+++ types/stripe-js/hosted-checkout.d.ts
